requirements.txt
playwright==1.58.0
behave
allure-behave
python-dotenv
requests